* commit '8a7489f45a532602bc9689a42f31de88545a44d2':
  Use long instead of int for address in DirectByteBuffer constructor

+++ luni/src/test/java/libcore/java/nio/BufferTest.java
@@ -715,7 +715,7 @@ public class BufferTest extends TestCase {
     public void testHasArrayOnJniDirectByteBuffer() throws Exception {
         // Simulate a call to JNI's NewDirectByteBuffer.
         Class<?> c = Class.forName("java.nio.DirectByteBuffer");
-        Constructor<?> ctor = c.getDeclaredConstructor(int.class, int.class);
+        Constructor<?> ctor = c.getDeclaredConstructor(long.class, int.class);
         ctor.setAccessible(true);
         ByteBuffer bb = (ByteBuffer) ctor.newInstance(0, 0);
